I learned about using pointers with arrays in C
it's also good to know that the C language actually doesn't require you to write a return statement for return-type methods (which is dangerous). If you do that any value would be returned, which would usually be the last value processed by your CPU (for me it's 48 since that was the last value in the array that I put in the method. It's good to know about this at least. This is what we call undefined behavior. TLDR: C doesn't show compile error if you forgot to write a return statement!